requirements.txt

streamlit>=1.35
google-cloud-firestore>=2.16.0
google-auth>=2.30.0
pandas>=2.2.2
pyarrow>=15.0.0
